+++ server/src/db/create_schema_and_tables.sql
@@ -35,6 +35,7 @@ CREATE TABLE `credentials` (
   `host` varchar(250) DEFAULT NULL,
   `port` int(11) DEFAULT NULL,
   `description` text NOT NULL,
+  `secure` tinyint(4) DEFAULT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `uk_AnsibleForms_credentials_natural_key` (`name`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;