+++ server/src/models/credential.model.js
@@ -16,6 +16,7 @@ var Credential=function(credential){
     this.host = credential.host;
     this.port = credential.port;
     this.user = credential.user;
+    this.secure = (credential.secure)?1:0;
     this.password = encrypt(credential.password);
     this.description = credential.description;
 };
@@ -79,7 +80,7 @@ Credential.findByName = function (name) {
   logger.debug(`Finding credential ${name}`)
   var cred = cache.get(name)
   if(cred==undefined){
-    return mysql.do("SELECT host,port,name,user,password FROM AnsibleForms.`credentials` WHERE name=?;",name)
+    return mysql.do("SELECT host,port,name,user,password,secure FROM AnsibleForms.`credentials` WHERE name=?;",name)
     .then((res)=>{
       if(res.length>0){
         res[0].multipleStatements = true